Feature/task11

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -17,33 +17,42 @@ import jp.co.seattle.library.service.BooksService;
 /**
  * 削除コントローラー
  */
-@Controller //APIの入り口
+@Controller // APIの入り口
 public class DeleteBookController {
-    final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
-    
-   
-  
-    @Autowired
-	private BooksService booksService;
-    /**
-     * 対象書籍を削除する
-     *
-     * @param locale ロケール情報
-     * @param bookId 書籍ID
-     * @param model モデル情報
-     * @return 遷移先画面名
-     */
-    @Transactional
-    @RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
-    public String deleteBook(
-            Locale locale,
-            @RequestParam("bookId") Integer bookId,
-            Model model) {
-        logger.info("Welcome delete! The client locale is {}.", locale);
-        booksService.deleteBookInfo(bookId);
-        model.addAttribute("bookList",booksService.getBookList());
-        
-        return "home";
+	final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
 
-    }
+	@Autowired
+	private BooksService booksService;
+
+	/**
+	 * 対象書籍を削除する
+	 *
+	 * @param locale ロケール情報
+	 * @param bookId 書籍ID
+	 * @param model  モデル情報
+	 * @return 遷移先画面名
+	 */
+	@Transactional
+	@RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
+	public String deleteBook(Locale locale, @RequestParam("bookId") Integer bookId, Model model) {
+		logger.info("Welcome delete! The client locale is {}.", locale);
+
+		int rentBookId = booksService.rent(bookId);
+
+		// 貸出されてなければ
+		if (rentBookId == bookId) {
+			model.addAttribute("lendingError", "貸し出し中なので削除できません。");
+			// 画面やり直し
+			model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+			return "details";
+
+		} else {
+			// 削除する
+			booksService.deleteBookInfo(bookId);
+			// 書籍一覧に戻る
+			model.addAttribute("bookList", booksService.getBookList());
+			return "home";
+			// 貸出中だったら
+		}
+	}
 }
